move CDP bridge setup to ConnectSession
