+++ packages/vscode-extension/src/connect/ConnectSession.ts
@@ -1,8 +1,13 @@
+import path from "path";
+import fs from "fs";
 import { Disposable } from "vscode";
 import { DebugSession, DebugSessionImpl } from "../debugging/DebugSession";
 import { Metro } from "../project/metro";
 import { BaseInspectorBridge } from "../project/bridge";
 import { disposeAll } from "../utilities/disposables";
+import { extensionContext } from "../utilities/extensionContext";
+import { Logger } from "../Logger";
+import { getTelemetryReporter } from "../utilities/telemetry";
 
 export interface ConnectSessionDelegate {
   onSessionTerminated: () => void;
@@ -72,10 +77,45 @@ export default class ConnectSession implements Disposable {
       this.debugSession.onBindingCalled((event: unknown) => {
         this.inspectorBridge.onBindingCalled(event);
       }),
+      this.debugSession.onBundleParsed(({ isMainBundle }) => {
+        if (isMainBundle) {
+          this.setupRadonConnectRuntime();
+        }
+      }),
     ];
     return new Disposable(() => {
       disposeAll(subscriptions);
     });
+  }
+
+  private async setupRadonConnectRuntime() {
+    try {
+      // load script from lib/connect_runtime.js and evaluate it
+      const runtimeScriptPath = path.join(
+        extensionContext.extensionPath,
+        "dist",
+        "connect_runtime.js"
+      );
+      const runtimeScript = await fs.promises.readFile(runtimeScriptPath, "utf8");
+
+      await this.debugSession.addBinding("__radon_binding");
+
+      const result = await this.debugSession.evaluateExpression({
+        expression: runtimeScript,
+      });
+      if (result.exceptionDetails) {
+        Logger.error("Failed to setup Radon Connect runtime", result.exceptionDetails);
+        getTelemetryReporter().sendTelemetryEvent("radon-connect:setup-runtime-error", {
+          error: result.exceptionDetails.exception?.description ?? "Unknown error",
+        });
+      }
+    } catch (e) {
+      const errorMessage = (e as Error).message;
+      Logger.error("Failed to setup Radon Connect runtime", e);
+      getTelemetryReporter().sendTelemetryEvent("radon-connect:setup-runtime-error", {
+        error: errorMessage || "Unknown error",
+      });
+    }
   }
 
   dispose() {
