+++ packages/vscode-extension/src/debugging/RadonCDPProxyDelegate.ts
@@ -1,5 +1,3 @@
-import path from "path";
-import fs from "fs/promises";
 import { IProtocolCommand, IProtocolSuccess, IProtocolError, Cdp } from "vscode-cdp-proxy";
 import { EventEmitter } from "vscode";
 import { Minimatch } from "minimatch";
@@ -7,8 +5,6 @@ import _ from "lodash";
 import { CDPProxyDelegate, ProxyTunnel } from "./CDPProxy";
 import { SourceMapsRegistry } from "./SourceMapsRegistry";
 import { Logger } from "../Logger";
-import { extensionContext } from "../utilities/extensionContext";
-import { getTelemetryReporter } from "../utilities/telemetry";
 
 export class RadonCDPProxyDelegate implements CDPProxyDelegate {
   private debuggerPausedEmitter = new EventEmitter<{ reason: "breakpoint" | "exception" }>();
@@ -55,7 +51,7 @@ export class RadonCDPProxyDelegate implements CDPProxyDelegate {
         return this.handleDebuggerResumed(applicationCommand, tunnel);
       }
       case "Debugger.scriptParsed": {
-        return this.handleScriptParsed(applicationCommand, tunnel);
+        return this.handleScriptParsed(applicationCommand);
       }
       case "Runtime.executionContextsCleared": {
         this.sourceMapRegistry.clearSourceMaps();
@@ -274,46 +270,13 @@ export class RadonCDPProxyDelegate implements CDPProxyDelegate {
     throw new Error("Source map URL schemas other than `data` and `http` are not supported");
   }
 
-  private async setupRadonConnectRuntime(tunnel: ProxyTunnel) {
-    // load script from lib/connect_runtime.js and evaluate it
-    const runtimeScriptPath = path.join(
-      extensionContext.extensionPath,
-      "dist",
-      "connect_runtime.js"
-    );
-    const runtimeScript = await fs.readFile(runtimeScriptPath, "utf8");
-
-    await tunnel.injectDebuggerCommand({
-      method: "Runtime.addBinding",
-      params: {
-        name: "__radon_binding",
-      },
-    });
-
-    const result = (await tunnel.injectDebuggerCommand({
-      method: "Runtime.evaluate",
-      params: {
-        expression: runtimeScript,
-      },
-    })) as Cdp.Runtime.EvaluateResult;
-    if (result.exceptionDetails) {
-      Logger.error("Failed to setup Radon Connect runtime", result.exceptionDetails);
-      getTelemetryReporter().sendTelemetryEvent("radon-connect:setup-runtime-error", {
-        error: result.exceptionDetails.exception?.description ?? "Unknown error",
-      });
-    }
-  }
-
   private handleBindingCalled(command: IProtocolCommand) {
     const params = command.params as Cdp.Runtime.BindingCalledEvent;
     this.bindingCalledEmitter.fire(params);
     return command;
   }
 
-  private async handleScriptParsed(
-    command: IProtocolCommand,
-    tunnel: ProxyTunnel
-  ): Promise<IProtocolCommand> {
+  private async handleScriptParsed(command: IProtocolCommand): Promise<IProtocolCommand> {
     const { sourceMapURL, url, scriptId } = command.params as Cdp.Debugger.ScriptParsedEvent;
     if (!sourceMapURL) {
       return command;
@@ -326,10 +289,6 @@ export class RadonCDPProxyDelegate implements CDPProxyDelegate {
       );
 
       this.sourceMapRegistry.registerSourceMap(sourceMapData, url, scriptId, isMainBundle);
-
-      if (isMainBundle && this.installConnectRuntime) {
-        await this.setupRadonConnectRuntime(tunnel);
-      }
 
       this.bundleParsedEmitter.fire({ isMainBundle });
     } catch (e) {
